Attempt at fixing an often failing test.
When trying to find if a candidate has already been selected, avoid choosing
a node that is in WAIT_PRIMARY. As we don't change the state of such a node
when it becomes unhealthy, that might lead to strange results when two
failover situations happen one after another.

Also guard demoted to cachingup transitions to only happen when the current
primary is known healthy.