Complete transition of a second (or greater) failed primary
Assume that there exists a healthy three datanode cluster having datanodes named
node1, node2, node3. Assume that node1 is the primary and it fails. The cluster
will correctly demote node1 and promote node2 to primary. When node2 fails, the
cluster would fail to reach a stable state due to node2 getting stuck in
draining state.

The culrpit is GetPrimaryOrDemotedNodeInGroup which contains logic to try to
find a demoted primary in case that a healty primary is not existing. To do
that, it traverses an ordered by name list of nodes in the group. When it finds
the first node that is either a demoted or draining or taken out for
maintainance primary, it chooses it and stops looking further. In the example
above, note that the ordered by name list starts with node1, which would satisfy
the criterion and node2 will never be selected.

To fix that, continue looping through the nodes. Assuming that there can be only
one primary active at any time, hence only one primary that is currently going
through demotion process, then choose that one.

Add a specific test case for it.