Make multi_alternate_primary_failures more resiliant to timing issues redoux
The reason of the flakiness seems to be at the end that the tests were waiting
for assigned state, so by the time of execution the transision might not have
been completed. Fix by adding waiting_for_state calls for those.

Also rename the *007* to *006*, no need to skip one.