Re-design retry mechanism due to a race in pre 12 Postgres versions
In PostgreSQL version prior to 12 it is possible to get erroneous results in
GetBackgroundWorkerPid due to the slot being out of date. This window becomes
more probable to hit, if the postmaster has been signalled. Versions greater or
equal to 12 seem to not be affected.

One could argue that a simple kill() command would suffice to see if the
bgworker process is still running or to use the SIGUSR1 mechanism provided by
bgworker.c to signal the parent processes when a worker starts or stops.
However, GetBackgroundWorkerPid is still the preferred and documented way to
retrieve status.

Now the code checks the result of GetBackgroundWorkerPid and
WaitForBackgroundWorkerStartup, and if it is not the expected, it removes the
entry from the worker hash table via StopHealthCheckWorker. That way, the code
still uses the recommended interfaces and makes certain that it always aware of
the proper state of each worker.

A distinct warn level message is emmited according to the PostgreSQL version.

Also remove a banned function, snprintf, by using StringInfo.