Refrain from early selection of a new primary.
If only one node is capable of being a primary and number_sync_standbys is
set to 1, it's better to leave the node in read-only in REPORT_LSN until we
have another node to be its secondary. Otherwise, write queries are going to
be blocked. With that approach, write queries are refused, which is
generally easier to deal with at the application level.