Fix async tests and allow wait_primary to be demoted at stop_replication.
When the new primary goes through the stop_replication state, then the old
primary needs to be demoted. In the async test suite, this now happens to a
node that might be wait_primary still, and is unhealthy. When this nodes
resumes and contact the monitor, it should not be wait_primary anymore.

We might want to revisit that idea and have a wait_primary node go directly
to either catchingup or maybe join_secondary in such as sequence.