Keep the context around for as long as you need it
A previous commit [1], rather embarrassingly, failed to notice the reset of used
memory contexts. Additionally it was missdiagnosed for PostgreSQL versions 10
and 11. A simple run with assertions enabled was enough to expose the real bug.

This commit ammends that error and in passing restructures a bit the code in
HealthCheckWorkerLauncherMain to be hopefully cleaner.

Now the launcherContext is still reset in every database scan, but objects that
need to persist, do reside in the parent context.

[1] commit 785fd533bb