An attempt to stabilize multi_async tests (#722)
It seems that node2.has_needed_replication_slots() requires to check
the postgres version. To do so, the first time such a check is required, a query
is executed on that node via self.run_sql_query(). At the time of the query the
node has been assigned and reached state secondary. However the cluster is not
yet in its final state, as node1 may still have to reach primary state.

In such a case, there exists a time window that the query will be executed while
node2 is consuming the change of node1.

By changing the order of the calls in that test, to call
has_needed_replication_slots() after node1 has finished the transition and the
cluster is in its stable state, the window is eliminated.

Note that an additional and probably superfluous after at the end of the
test_002 to assert the state of primary is maintained in accordance to the
pattern seen in other tests in the same suite. (e.g. test_003_add_standby,
test_009_add_sync_standby, etc).

Co-authored-by: Georgios Kokolatos <gkokolatos@protonmail.com>