Fix how many nodes need to report their LSN to perform a failover. (#707)
* Fix how many nodes need to report their LSN to perform a failover.

With this patch we make sure that all the nodes that participate in the
quorum have reported their LSN before we select a candidate for failover.
Otherwise, we might lose committed data.

* Refrain from early selection of a new primary.

If only one node is capable of being a primary and number_sync_standbys is
set to 1, it's better to leave the node in read-only in REPORT_LSN until we
have another node to be its secondary. Otherwise, write queries are going to
be blocked. With that approach, write queries are refused, which is
generally easier to deal with at the application level.

* Fix async tests and allow wait_primary to be demoted at stop_replication.

When the new primary goes through the stop_replication state, then the old
primary needs to be demoted. In the async test suite, this now happens to a
node that might be wait_primary still, and is unhealthy. When this nodes
resumes and contact the monitor, it should not be wait_primary anymore.

We might want to revisit that idea and have a wait_primary node go directly
to either catchingup or maybe join_secondary in such as sequence.

* Per review, we should have nodes in the quorum to allow failover.

Co-authored-by: Jelte Fennema <github-tech@jeltef.nl>