Implement support for citus.local_hostname. (#731)
* Implement support for citus.local_hostname.

* Allow retry when configuration file does not exists.

There seem to be a race condition at process startup where we overwrite the
configuration file when merging the command line arguments with the
pre-existing setup. For a short while the configuration file is reported
missing (ENOENT).

Implement a very small retry loop so that `pg_autoctl do pgsetup wait` is
more reliable, in particular in the test suite.