Allow pg_autoctl drop node to finish removing a lost node.
When a node won't come back up again, the first call to pg_autoctl drop node
sets the assigned role to DROPPED, but the node is not in a position to call
node_active() to clean this entry.

Now, another call to pg_autoctl drop node from the monitor allows to
clean-up the node for real.

If the node still comes back up again, the situation is properly detected
and the command pg_autoctl create node --run is now able to re-register the
node with its old nodeid and continue from there.