Implement drop-at-a-distance semantics. (#734)
* Implement drop-at-a-distance semantics.

This allows running `pg_autoctl drop node --name foo` from the monitor, or
even using the SQL API directly, and have the node realise it's been
dropped. And then stop.

The configuration file, state file, and PGDATA are not touched, for that the
command pg_autoctl drop node --pgdata ... --destroy can be used on the node
itself.

* Allow pg_autoctl create node on top of a dropped node.

* Allow pg_autoctl drop node to finish removing a lost node.

When a node won't come back up again, the first call to pg_autoctl drop node
sets the assigned role to DROPPED, but the node is not in a position to call
node_active() to clean this entry.

Now, another call to pg_autoctl drop node from the monitor allows to
clean-up the node for real, with the force option set to true.

If the node still comes back up again, the situation is properly detected
and the command pg_autoctl create node --run is now able to re-register the
node with its old nodeid and continue from there.

* Per review, introduce pg_autoctl drop node --force.

* Per review, fix some pg_autoctl drop node cases.

In particular dropping a local node that is running in the background was
broken. Refactor and simplify the code to make all cases work:

  1. stop local node and drop
  2. drop local node while it's running in the background
  3. drop node from the monitor while it's running
  4. drop node from the monitor after having stopped it

In case 4. we need to then resort to pg_autoctl drop node --force, as expected.

* Allow Postgres 14 failures.

* Improve handling of node being drop during init phases, per review.

* Per review, allow re-creating a primary node from a dropped node.