Fix transaction begin failure handling
When opening a transaction using `pgsql_begin` we would set the
connection type to multi statement mode. However, if we failed to open a
connection, we would not reset it to single statement mode.

This changes it such that we only set the connection type to multi
statement mode once we have successfully opened a transaction.

See #746