Set candidate priority without waiting for state changes in more cases.
When dropping the last node with candidate priority greater than zero, we
might want to change the candidate priority of the last remaining node. The
client-side implementation of the command would wait for the APPLY_SETTINGS
cycle in some cases with dropped nodes, which is unfortunate: this PR fixes
that.