Avoid re-electing primary during a switchover.
When all the nodes are up and running, including the primary, a switchover
has a great chance of finding that the current primary is then the best
candidate to with the primary election, making the operation a no-op.

In this PR we tweak the current primary node's candidate priority in a way
that should make it lose the election, and reset the priority when the
election is over.