Fix connection retry when ping succeeds and then connection fail.
This happens when e.g. the HBA rules have not been set yet on the monitor.
We have a timing where we could ping successfully and then fail to connect,
and in that situation we had a spurious call to pgsql_finish() that would
reset the connection statement type to SINGLE, even when it was before set
to SINGLE.