Fix replication slot maintenance on secondary nodes. (#781)
It might happen that a node continues streaming from a secondary after a
failover, and in that case the streaming replication protocol takes care of
maintaining the replication slot. Trying to use pg_replication_slot_advance
would then fail.

This situation currently can arise when a node is in maintenance a failover
happens, where the node in maintenance stays connected to the old-primary,
now repurposed as a secondary.

This situation could also arise when a cascading replication setup is
considered. We don't have direct support for those architectures at the
moment, though that's something of interest for the future.