Improve all diagrams.
Keep the concept of the first diagram with the kind of a circle with four
nodes (application, primary, secondary, monitor) and build on-top of the
same idea for the others diagrams.

Every diagram builds by adding some level of complexity on-top of the
previous one. By keeping a central theme it makes it obvious and much easier
to re-use the understanding from the previous diagram in the next.