Refrain from DRAINING state when there is no candidate.
When a primary is not healthy and there is no candidate node to failover to,
then assigning to the primary the DRAINING state is not helping.

Also, the reason why we don't have a candidate at the moment might be that
the other nodes are in MAINTENANCE and the operator is restarting Postgres
to install some new configuration.