Per review, enhance multi-standby maintenance steps for a primary.
When doing pg_autoctl enable maintenance --allow-failover on a primary node
of a multiple standbys system, we can bypass the usual prepare_maintenance
step by putting the other nodes in report_lsn immediately to trigger an
election.

Doing it that way has the benfits also of allowing the primary to reach
maintenance while the election is happening, which in turns allow the
operator to change their mind with pg_autoctl disable maintenance on the
former primary node.