Fix problems with bad migration to 1.5
The script to migrate from 1.4 to 1.5 did not correctly remove the old
`register_node` function. This caused issues for the migration from 1.5
to 1.6, because we remove the old `replication_state` type there. The
old `register_node` function still depended on this version of the
`replication_state` type, so it's removal was blocked.