Get rid of the JOIN_PRIMARY state.
We still need to maintain backward compatibility of course, in case there is
an upgrade where one node is in JOIN_PRIMARY and then we want to be able to
continue working with the node.

Now, any node that reports JOIN_PRIMARY is unconditionnally assigned PRIMARY
instead, and we never assign JOIN_PRIMARY anymore. We still have some code
that matches the state in the group level FSM, and we keep the full compat
on the client side for the moment.