Per valgrind, fix some memory leaks.
We would be using PQfreemem on a PQExpBuffer, where it's expected to be
using destroyPQExpBuffer instead.

Also we were short of a call to finish the
connection to the monitor notification client when exiting from the service
keeper process, although adding this clean-up doesn't seem to impact the
valgrind reports.