Make sure to disable sync rep when initializing a primary. (#801)
The step to reach the SINGLE state could be done from a pre-existing PGDATA,
such as taken from a backup tool. In that case, the postgresql.auto.conf
might have been installed with a setting that enables synchronous
replication.

Trouble is, we're initializing the first node in the system, and when we
reach the SINGLE state any setup for sync-rep does not make sense, we know
we don't have any secondary available.