Work around pg_replication_slot_advance xmin maintenance bug.
When the Postgres function pg_replication_slot_advance is called on a slot
that used to be active (using streaming replication) and is now
inactive (maintained manually) then the xmin of the slot is not maintained
anymore.

To avoid situations where Postgres won't VACUUM dead rows and otherwise
maintains itself correctly, when a former primary node is available again
and joind a pg_auto_failover as a standby, we drop the pre-existing
replication slot and create new ones (where this time xmin is NULL).