Relax the checks/assertions on which PID owns the semaphore
It turns out that the checks to ensure which PID owns the semaphore
is too aggressive. The PID might be 0, pid of our own process or
any process above to our PID in the process tree.

So, we simplify the rule. We only terminate the semaphore when
we are owner of the PID. Else, we wait until the owner of the
PID exits (which will eventually happen).