Make Dockerfile aware of Postgres versions.
That way we can easily build and test pg_auto_failover locally with any
supported version of Postgres. We use docker build --build-arg for that, and
split the Dockerfile in more stages for clarity and better cache re-use.

It is now possible to run the same test for Postgres 10 then 14 that way:

    $ make run-test TEST=test_auth PGVERSION=10
    $ make run-test TEST=test_auth PGVERSION=14