Add chapters to the docs navigation, allow building docs in docker.
We have lots of documentation and the main navigation is crowded, so split
it in several chapters with clear names.

Also, allow to build (and serve) the docs from a docker image to make it
easy for contributors to check their work. This also makes it possible to
build the docs in the CI.