Use environment PGDATABASE as the default create node --dbname. (#956)
When creating a node the option --dbname can be used to create a database
for the application. When using Citus, that database is where pg_autoctl
installs the Citus extensions, which is recommanded to be installed in a
single database in your Postgres cluster.

To ease using pg_autoctl docker-compose and kubernetes, use PGDATABASE as
the default value for the --dbname option, when set in the environment.