Fix pg_autoctl monitor bad memory access
The current code in pg_autoctl/monitor.c causes a use-after-scope bad
memory access because it assigns a pointer inside a block that is then
used outside this scope.

The fix is to move the initialization of the referenced value to the
beginning of the function, outside the block scope.